pyproject.toml
[tool.poetry]
name = "dpo"
version = "0.1.0"
description = "dpo dataset on huggingface"
authors = ["Shun-Fu Hu"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
openpyxl = "^3.1.2"
transformers = "^4.38.2"
torchtext = "^0.17.1"
torch = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
